pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "paper-reader"
version = "0.0.0"
description = "Convert academic PDFs to LLM-ready markdown with chunked RAG output"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
dependencies = [
    "marker-pdf",
    "mcp[cli]",
    "click",
]

[project.scripts]
paper-reader = "paper_reader.cli:main"
paper-reader-mcp = "paper_reader.server:main"

[tool.hatch.build.targets.wheel]
packages = ["src/paper_reader"]
